"Hello" Using xlwings
This is a simple project which explores the use of Excel as a frontend part of an application, or graphical user interface (GUI), for a python program. It uses the Excel add-in xlwings as a tool for connecting the hello.py file to the hello.xlsm file in a seamless manner, with the RunPython function in VBA acting as the mediator between the two files. An interactive button in the Hello.xlsm file can be clicked to run the python script.

Have xlwings enabled as a reference for this project. Also ensure that the pythonpath in the xlwings ribbon is set to the path of your python script.